+++ js/scripts/JSRoot.menu.js
@@ -61,6 +61,10 @@ JSROOT.define(['d3', 'jquery', 'painter', 'jquery-ui'], (d3, $, jsrp) => {
          this.remove_bind = this.remove.bind(this);
       }
 
+      /** @summary Returns object with mouse event position when context menu was actiavted
+        * @desc Return object will have members "clientX" and "clientY" */
+      getEventPosition() { return this.show_evnt; }
+
       /** @summary Add menu item
         * @param {string} name - item name
         * @param {function} func - func called when item is selected */